more wombat fixes
